Settings: Do not allow to access quick settings on secure lockscreen - option (2/2)
We have a lot security guys as user due of using the phones as business phones...that
said a lot complains go into the direction of google how it is possible to have
access to quick settings on a secure lockscreen which is absolutely understandable
due that you are able to control Wifi BT AirplaneMode data etc.

This patch takes care on this problem and adds an option to control this behaviour.

ATTENTION: Due that this is really an UX derp for an average user this feature is
enabled by default now. User need to disable it explicit in settings to get on secure
locksreen the original android behaviour back.

Change-Id: Iac8c181767a16d3d198672d16901670984b58f48

original commit: https://github.com/SlimRoms/packages_apps_Settings/commit/a23400c0cfb37171cf547715a3bf04c24f812a28